% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldSeek.R
\name{available_databases}
\alias{available_databases}
\title{Displays databases available in FoldSeek.}
\usage{
available_databases()
}
\value{
A list with headers name, version, and path corresponding to each
available database's name, version, and path for the search query.
}
\description{
Displays the available databases for the FoldSeek query.
}
\examples{
# Example 1
available_databases()

}
