% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clears the cache for an accession.}
\usage{
clear_cache(accession)
}
\arguments{
\item{accession}{The Uniprot Accession associated with the storage location.}
}
\description{
Clears the cache for the Uniprot accession in which the FoldSeek similarity
search was conducted on.
}
\examples{
# Example 1
clear_cache("P00520")

}
