% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldSeek.R
\name{foldseek}
\alias{foldseek}
\title{Queries the given Uniprot accession or FASTA file via FoldSeek and returns
the top predictions for closest similarity in structure.}
\usage{
foldseek(target, k)
}
\arguments{
\item{target}{A FASTA string or FASTA filepath, or a valid Uniprot Accession
number (numeric or string).}

\item{k}{The top k number of entries most similar to the file are returned.}
}
\value{
The k-JSON bodies for each of the resulting outputs from the query.
}
\description{
Queries the given Uniprot accession or FASTA file via FoldSeek and returns
the top predictions for closest similarity in structure.
}
\examples{
# Example 1
foldseek(target = "./test.fasta", k = 3)

# Example 2
foldseek(target = "P00520", k = 5)

}
