% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaFold.R
\name{pull_prediction}
\alias{pull_prediction}
\title{Pulls AlphaFold Prediction.}
\usage{
pull_prediction(accession)
}
\arguments{
\item{accession}{A Uniprot Accession string.}
}
\value{
A JSON body containing the predicted protein architecture including
entries for gene, pdbUrl, cifUrl, paeImageUrl, and uniprotSequence.
}
\description{
Pulls the AlphaFold protein structure prediction information.
}
\examples{
# Example 1
pull_prediction("P00520")

}
