% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaFold.R
\name{pull_summary}
\alias{pull_summary}
\title{Pulls the AlphaFold protein structure summary information.}
\usage{
pull_summary(qualifier)
}
\arguments{
\item{qualifier}{A Uniprot Accession number (numeric or string).}
}
\value{
The JSON body returned as a list whereby summary contains two
headers: $structures and $uniprot_entry. These headers contain information
about the structure produced (including urls on where to find them) and
information about the Uniprot accession entry respectively.
}
\description{
Pulls the AlphaFold protein structure summary information.
}
\examples{
# Example 1
pull_summary(qualifier = "P00520")

}
