% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaFold.R
\name{pull_url}
\alias{pull_url}
\title{Pulls the URL for the protein.}
\usage{
pull_url(accession, url_type)
}
\arguments{
\item{accession}{A Uniprot Accession number (numeric or string).}

\item{url_type}{The file type for the file associated with the url.}
}
\value{
The url for the predicted protein architecture file of the
corresponding file type.
}
\description{
Pulls the AlphaFold protein structure prediction URL for the protein
associated with the Uniprot Accession.
}
\examples{
# Example 1
pull_url("P00520", "pdb")

# Example 2
pull_url("P00520", "cif")

}
