% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldSeek.R
\name{status}
\alias{status}
\title{Gets status of job.}
\usage{
status(ticket_id)
}
\arguments{
\item{ticket_id}{A ticket ID provided by the job submitted.}
}
\value{
Status update corresponding to the ticket (ie. "PENDING", "ERROR",
or "COMPLETE").
}
\description{
Returns status of the job for the specified ticket ID.
}
\examples{
# Example 1
status("G04AJT4Nl6Pk-o8VzLYklMyOKADAIZwUJvpIyA")

}
