% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldSeek.R
\name{ticket}
\alias{ticket}
\title{Submits a job.}
\usage{
ticket(filepath)
}
\arguments{
\item{filepath}{A filepath to a CIF or PDB file, or a string containing the
information in said file.}
}
\value{
Ticket ID corresponding to submitted job.
}
\description{
Submits a job to the FoldSeek Search Server via API against all available
databases.
}
\examples{
# Example 1
ticket("./test.cif")

# Example 2
ticket("./test.pdb")

}
