% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeFold.R
\name{visualize_prediction}
\alias{visualize_prediction}
\title{Produces a visual of the AlphaFold prediction for this Uniprot accession.}
\usage{
visualize_prediction(qualifier)
}
\arguments{
\item{qualifier}{A Uniprot Accession number (numeric or string).}
}
\description{
Produces a visual of the AlphaFold prediction for this Uniprot accession.
}
\examples{
# Example 1
visualize_prediction(qualifier = "P00520")

}
