% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_top_hits}
\alias{cache_top_hits}
\title{Caches top hits for a job.}
\usage{
cache_top_hits(job, k, database, accession)
}
\arguments{
\item{job}{A job run via the foldseek() function.}

\item{k}{Number of top hits to save for visualization.}

\item{database}{The database to pull the top hits from. Defaults to the first
database in the job.}
}
\value{
.
}
\description{
Saves the top hits to cache for visualization.
}
\examples{
# Example 1
job <- foldseek("./path/to/test.cif")
cache_top_hits(job, 3)

}
