% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldSeek.R
\name{foldseek}
\alias{foldseek}
\title{Runs a FoldSeek job.}
\usage{
foldseek(filepath)
}
\arguments{
\item{filepath}{A  filepath to a valid CIF, mmCIF, or PDB file of a
protein of interest.}
}
\value{
A job data frame of the format

\if{html}{\out{<div class="sourceCode">}}\preformatted{$queries
  $header : <target protein header>
  $sequence : <target protein sequence>
$mode : <search mode>
$results :
  $db : <databases>
  $alignments : <hits for each database>
    $query : <job>
    $target : <hit protein name>
    $seqId : <sequence identity>
    $alnLength : <alignment length>
    $missmatches : <number of mismatches>
    $gapsopened : <number of gaps>
    $qStartPos : <>
    $qEndPos : <>
    $dbStartPos : <>
    $dbEndPos : <>
    $prob : <>
    $eval : <>
    $score : <>
    $qLen : <>
    $aAln : <alignment q score>
    $tCa : <>
    $tSeq : <>
    $taxId : <taxonomy id>
    $taxName : <taxonomy name>```
}\if{html}{\out{</div>}}
}
\description{
Queries the given CIF, mmCIF or PDB file via FoldSeek and returns
the job result for closest similarity in structure. This handles the ticket
and status portion.
}
\examples{
# Example 1
foldseek("./data/test.cif")

}
