% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldSeek.R
\name{foldseek}
\alias{foldseek}
\title{Runs a FoldSeek job.}
\usage{
foldseek(filepath)
}
\arguments{
\item{filepath}{A filepath to a valid CIF, mmCIF, or PCB file of a protein
of interest.}
}
\value{
A job data frame of the format

$queries
$header : \if{html}{\out{<target protein header>}}
$sequence : \if{html}{\out{<target protein sequence>}}
$mode : \if{html}{\out{<search mode>}}
$results :
$db : \if{html}{\out{<databases>}}
$alignments : \if{html}{\out{<hits for each database>}}
$query : \if{html}{\out{<job>}}
$target : \if{html}{\out{<hit protein name>}}
$seqId : \if{html}{\out{<sequence identity>}}
$alnLength : \if{html}{\out{<alignment length>}}
$missmatches : \if{html}{\out{<number of mismatches>}}
$gapsopened : \if{html}{\out{<number of gaps>}}
$qStartPos : <>
$qEndPos : <>
$dbStartPos : <>
$dbEndPos : <>
$prob : <>
$eval : <>
$score : <>
$qLen : <>
$aAln : \if{html}{\out{<alignment q score>}}
$tCa :
$tSeq :
$taxId : \if{html}{\out{<taxonomy id>}}
$taxName : \if{html}{\out{<taxonomy name>}}
}
\description{
Queries the given CIF, mmCIF or PDB file via FoldSeek and returns
the job result for closest similarity in structure.
}
\examples{
# Example 1
foldseek("./path/to/test.cif")

}
