% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaFold.R
\name{pull_prediction}
\alias{pull_prediction}
\title{Pulls the AlphaFold protein structure prediction information.}
\usage{
pull_prediction(accession)
}
\arguments{
\item{accession}{A Uniprot Accession number (numeric or string).}
}
\value{
The JSON body returned as a list whereby prediction[\link{1}] contains
the object for the predicted protein architecture. Notably, prediction[\link{1}]
contains entries for entryId, gene, pdbUrl, cifUrl, paeImageUrl, and
uniprotSequence.
}
\description{
Pulls the AlphaFold protein structure prediction information.
}
\examples{
# Example 1
pull_prediction("P00520")

}
