% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaFold.R
\name{pull_summary}
\alias{pull_summary}
\title{Pulls AlphaFold Summary.}
\usage{
pull_summary(accession)
}
\arguments{
\item{accession}{A Uniprot Accession string.}
}
\value{
A JSON body containing the predicted protein architecture under
two headers: \code{structures} and \code{uniprot_entry}.
}
\description{
Pulls the AlphaFold protein structure summary information.
}
\examples{
# Example 1
pull_summary("P00520")

}
