% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeFold.R
\name{visualize_foldseeks}
\alias{visualize_foldseeks}
\title{Produces a visual of the AlphaFold prediction and the closest FoldSeek
predictions for this Uniprot accession.}
\usage{
visualize_foldseeks(qualifier)
}
\arguments{
\item{qualifier}{A Uniprot Accession number (numeric or string).}
}
\description{
Produces a visual of the AlphaFold prediction and the closest FoldSeek
predictions for this Uniprot accession.
}
\examples{
# Example 1
visualize_foldseeks(qualifier = "P00520")

}
