% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeFold.R
\name{visualize_foldseeks}
\alias{visualize_foldseeks}
\title{Visualizes FoldSeek Predictions.}
\usage{
visualize_foldseeks(accession)
}
\arguments{
\item{accession}{A Uniprot Accession number (numeric or string).}
}
\description{
Produces a visual of the AlphaFold prediction and the closest FoldSeek
predictions for this accession.
}
\examples{
# Example 1
visualize_foldseeks("P00520", 3)

}
